Comments!!!
Documentation was added throughout the entire app (well, not as much in index.html, tbh). Didn't update the ReadMe, which is probably way more important for this particular project, but I want to develop good habits of documenting my code, so that anyone/myself can pick up right where I left off, if that is ever necessary.